Return type updated from u64 to usize

+++ src/interface/input.rs
@@ -6,7 +6,7 @@ pub fn input_data() -> String {
     data.trim().to_string()
 }
 
-pub fn input_data_int() -> i32 {
+pub fn input_data_int() -> usize {
     loop {
         let mut data: String = String::new();
         io::stdin().read_line(&mut data).expect("Input failure...");
